\name{rrdflibs-package}
\alias{rrdflibs-package}
\alias{rrdflibs}
\docType{package}
\title{
Jena RDF libraries to work with RDF data.
}
\description{
Package with the Jena library jars, mostly aimed for use
in the rrdf package. Using the rjava package, the Jena libraries
can also be used directly.
}
\details{
\tabular{ll}{
Package: \tab rrdflibs\cr
Type: \tab Package\cr
Version: \tab 1.1.2\cr
Date: \tab 2011-04-25\cr
License: \tab AGPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Egon Willighagen

Maintainer: Egon Willighagen <egon.willighagen@gmail.com>
}
\keyword{ package }
\keyword{ RDF }
\keyword{ SPARQL }
